pyproject.toml
[project]
name = "mindrtukov-scripts"
version = "0.1.0"
description = "Telegram archiver and utilities for mindrtukov"
requires-python = ">=3.9"

[project.optional-dependencies]
all = [
  "telethon>=1.41.0",
  "python-dotenv>=1.0.1",
]

[tool]
